feat(lifecycle): generalize ExitGuard
Add a generic ExitGuard to the lifecycle crate to avoid ad-hoc
definitions.
